classifier/requirements.txt
tensorflow-addons
jupyter
